requirements.txt
streamlit
matplotlib
pandas
st_pages
Pillow
requests
st-gsheets-connection
koreanize_matplotlib